extract miner from watch-donations
